udpate env variables and Basic auth
